Jarvis/docker_tts/requirements.txt
azure-cognitiveservices-speech==1.25.0  
fastapi==0.115.0
uvicorn==0.31.1
flask==2.3.2
gunicorn==20.1.0
requests==2.31.0
numpy==1.25.2
typer==0.9.0
nvidia-modelopt==0.11.1
pynvml==11.5.0
loguru==0.7.0
python-dotenv==1.0.1